--- a/pkg/fs/rwosfs_test.go
+++ b/pkg/fs/rwosfs_test.go
@@ -181,7 +181,7 @@ func TestDirFSConsistentOrdering(t *testing.T) {
 	var results []string
 	for i := 0; i < 10; i++ {
 		var result []string
-		err := fs.WalkDir(fsys, "/", func(path string, d fs.DirEntry, err error) error {
+		err := fs.WalkDir(fsys, "/", func(path string, _ fs.DirEntry, err error) error {
 			require.NoError(t, err)
 			result = append(result, path)
 			return nil