Previously this relied on an io.ReadSeekCloser opener. That style of access induced a syscall every time we opened a file, which ends up being pretty slow if we access a ton of files.

Now, this takes an already open io.ReaderAt, which means we only call open a single time.

Separately, tarfs will now pre-compute results for ReadDir. We used to iterate over every file and check to see if its parent matched the argument. Now, we just compute all the results up front with a much more efficient mechanism that pre-allocates everything and does a sorted insertion (instead of allocating and sorting in every ReadDir call).

Looking at a CPU profile of running "melange scan" on sqlpad...

```
     open(): 145ms ->  0ms
sca.Analyze: 275ms -> 20ms
  tarfs.New:  68ms -> 72ms
```

We pay a very small up-front cost in tarfs.New in exchange for much faster file access, especially for ReadDir.